requirements.txt
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
gdown==4.5.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
nano==0.10.0
numpy==1.21.6
opencv-python==4.6.0.66
Pillow==9.2.0
protobuf==3.20.1
PySocks==1.7.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tensorboardX==2.5.1
torch==1.11.0+cu113
torchaudio==0.12.1
torchgeometry==0.1.2
torchvision==0.12.0+cu113
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.12
utils==1.0.1
Werkzeug==2.2.2
zipp==3.8.1
